Refactor Layout components to be app Root
