--- src/client/pages/index.ts
@@ -1,5 +1,0 @@
-import Home from './Home';
-import PlantIndex from './PlantIndex';
-import Quiz from './Quiz';
-
-export { Home, PlantIndex, Quiz };
